govwin-ingest/requirements.txt
azure-core==1.34.0
azure-cosmos==4.9.0
azure-functions==1.23.0
azure-identity==1.23.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
idna==3.10
MarkupSafe==3.0.2
msal==1.32.3
msal-extensions==1.3.1
pycparser==2.22
PyJWT==2.10.1
requests==2.32.4
six==1.17.0
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3
